Refactoring. Increasing mutation id size from integer to long.

+++ Unite.Data/Entities/Mutations/MutationOccurrence.cs
@@ -2,9 +2,9 @@
 {
     public class MutationOccurrence
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int AnalysedSampleId { get; set; }
-        public int MutationId { get; set; }
+        public long MutationId { get; set; }
 
         public virtual AnalysedSample AnalysedSample { get; set; }
         public virtual Mutation Mutation { get; set; }
